Deprecated functions in package datawrangling.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("<function>-deprecated").

code.categorical(x, variable = "", order = c())

duplicates.remove(x, timing.variables = "SessionStartDateTimeUtc",
  subj.variable = "Subject", taskname = "", output.folder = "")

files.bind(path = "", pattern = "", delim = "\t", na = "",
  output.file = "", bind = "rows")

files.join(path = "", pattern = "", delim = "\t", na = "",
  output.file = "", id = NULL)

files.merge(path = "", pattern = "", delim = "\t", na = "",
  output.file = "", id = NULL)

remove.latent(x, factor.list, missing.allowed, id, output.dir = NULL,
  output.file = NULL)

remove.save(x, remove, output.dir = NULL, output.file = NULL,
  id = "Subject")

reshape.gather(x, variable.names, values, id = NULL,
  separate.pattern = NULL)

reshape.spread(x, variables = "", variables_combine.sep = "_",
  values = c(), id = "", fill = NA)

code.categorical

For code.categorical, use code_categorical.

duplicates.remove

For duplicates.remove, use duplicates_remove.

files.bind

For files.bind, use files_bind.

files.join

For files.join, use files_join.

files.merge

For files.merge, use files_join.

remove.latent

For remove.latent, use remove_latent.

remove.save

For remove.save, use remove_save.

reshape.gather

For reshape.gather, use reshape_gather.

reshape.spread

For reshape.spread, use reshape_spread.

Examples

code.categorical(x, variable = "columnName", order = c("string1", "string2", "string3"))

#> Warning: 'code.categorical' is deprecated.
#> Use 'code_categorical' instead.
#> See help("Deprecated")

#> Error in x[which(x[variable] == order[value]), variable] <- value: object 'x' not found

duplicates.remove(x, subset = c(colnames(x)), context = "Subject")

#> Error in duplicates.remove(x, subset = c(colnames(x)), context = "Subject"): unused arguments (subset = c(colnames(x)), context = "Subject")

files.bind(path = "./Data", pattern = ".txt")

#> Warning: 'files.bind' is deprecated.
#> Use 'files_bind' instead.
#> See help("Deprecated")

#> # A tibble: 0 x 0

files.join(path = "./Data", pattern = ".txt")

#> Warning: 'files.join' is deprecated.
#> Use 'files_join' instead.
#> See help("Deprecated")

#> Error in dfs[[1]]: subscript out of bounds

files.merge(path = "./Data", pattern = ".txt")

#> Warning: 'files.merge' is deprecated.
#> Use 'files_join' instead.
#> See help("Deprecated")

#> Error in dfs[[1]]: subscript out of bounds

# Example
remove.save(data, remove = data_remove, save = "data/remove", taskname = "Flanker")

#> Error in remove.save(data, remove = data_remove, save = "data/remove",     taskname = "Flanker"): unused arguments (save = "data/remove", taskname = "Flanker")

reshape.gather(x, variables = "variable", values = c("value1", "value2"), by = "Subject")

#> Error in reshape.gather(x, variables = "variable", values = c("value1",     "value2"), by = "Subject"): unused arguments (variables = "variable", by = "Subject")

reshape.spread(x, variables = "variable", values = c("value1", "value2"), by = "Subject")

#> Error in reshape.spread(x, variables = "variable", values = c("value1",     "value2"), by = "Subject"): unused argument (by = "Subject")